Change name as SwapWidget
